cloud_cowork/requirements.txt
# Local JARVIS Desktop GUI Dependencies (No External APIs)
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.28.0
aiohttp>=3.8.0
pycaw>=2023.3.1
screen-brightness-control>=0.13.2
pywhatkit>=5.4.0
pyautogui>=0.9.54
keyboard>=0.13.5
ollama>=0.1.7
pyttsx3>=2.2.0
speechrecognition>=3.10.0
pyaudio>=0.2.11
